requirements.txt
mutalyzer-algebra
requests
dash-cytoscape